Enhance patient management and UI components
- Updated `addPatient_view.dart` to improve patient data handling and image upload functionality.
- Enhanced `help.dart` with better UI and search functionality in the help view.
- Improved `login_view.dart` with better login handling and user feedback.
- Updated `miniPatientCard.dart` to handle dynamic disease summaries and improved dismissible functionality.
- Enhanced `miniProfileCard.dart` with better profile image handling.
- Updated `pallet.dart` with new color constants and text styles.
- Improved `patientProfile_view.dart` with better patient data handling and UI improvements.
- Enhanced `patientsList.dart` with better patient list handling and deletion confirmation.
- Updated `profile.dart` with better user profile management.
- Improved `search_view.dart` with better patient search and filtering functionality.
- Updated `textfield.dart` with better text field styles and functionalities.
- Updated `pubspec.yaml` with new dependencies and asset paths.
- Enhanced `profileData.dart` with better profile image and user data handling.
- Improved `splashScreen.dart` with better splash screen animation handling.

+++ lib/widgets/patientsList.dart
@@ -9,7 +9,7 @@ class PatientList extends StatelessWidget {
   PatientList({super.key, required this.patients});
 
   @override
-  Widget build(BuildContext context) {
+  Widget build(BuildContext maincontext) {
     return ListView.builder(
       itemCount: patients.length,
       itemBuilder: (context, index) {
@@ -38,16 +38,52 @@ class PatientList extends StatelessWidget {
           phoneNumber: patient['phoneNumber'] ?? 'N/A',
           disease: diagnosisSummary,
           onDismissed: () {
-            // Perform deletion
-            _auth.deletePatient(patient['id']);
+            // Show confirmation dialog before deletion
+            showDialog(
+              context: context,
+              builder: (BuildContext context) {
+                return AlertDialog(
+                  title: Text('Confirm Deletion'),
+                  content: Text(
+                      'Are you sure you want to delete patient ${patient['firstName']} ${patient['lastName']}?'),
+                  actions: [
+                    TextButton(
+                      onPressed: () {
+                        Navigator.of(context).pop(); // Close the dialog
+                      },
+                      child: Text('Cancel'),
+                    ),
+                    TextButton(
+                      onPressed: () {
+                        Navigator.of(context).pop(); // Close the dialog
 
-            // Show feedback
-            ScaffoldMessenger.of(context).showSnackBar(
-              SnackBar(
-                content: Text(
-                    "Patient ${patient['firstName']} ${patient['lastName']} deleted."),
-                duration: Duration(seconds: 1),
-              ),
+                        // Show a snackbar to indicate deletion with undo option
+                        final snackBar = SnackBar(
+                          content: Text(
+                              'Patient ${patient['firstName']} ${patient['lastName']} deleted'),
+                          action: SnackBarAction(
+                            label: 'Undo',
+                            onPressed: () {},
+                          ),
+                        );
+
+                        // Show the SnackBar and wait for it to close
+                        ScaffoldMessenger.of(maincontext)
+                            .showSnackBar(snackBar)
+                            .closed
+                            .then((SnackBarClosedReason reason) {
+                          if (reason != SnackBarClosedReason.action) {
+                            // Call the deletion function if not undone
+                            _auth.deletePatient(patient['id']);
+                          }
+                        });
+                      },
+                      child:
+                          Text('Delete', style: TextStyle(color: Colors.red)),
+                    ),
+                  ],
+                );
+              },
             );
           },
         );
